rel pred loss

+++ Mindblocks/default_component_types/neural_network/cross_entropy.py
@@ -17,7 +17,7 @@ class CrossEntropy(ComponentTypeModel):
 
     def execute(self, input_dictionary, value, output_value_models, mode):
         cross_entropy = tf.nn.sparse_softmax_cross_entropy_with_logits(
-                labels=tf.squeeze(input_dictionary["labels"].get_value()),
+                labels=tf.cast(tf.squeeze(input_dictionary["labels"].get_value()), tf.int32),
                 logits=input_dictionary["logits"].get_value())
         output_value_models["output"].assign(cross_entropy)
         return output_value_models
